refactor: 절대경로 적용
