+++ src/shared/assets/SVGicons/32pxIcon.jsx
